chore: Move slider design to `Slider.tsx` since we came up with a "better" name
